feat(docs): extract swagger into a swagger.setup.ts file
