+++ src/main.ts
@@ -1,20 +1,13 @@
 import { ValidationPipe } from '@nestjs/common'
 import { NestFactory } from '@nestjs/core'
-import { SwaggerModule, DocumentBuilder } from '@nestjs/swagger'
 import { AppModule } from './app.module'
+import { setupSwagger } from './swagger.setup'
 
 async function bootstrap(): Promise<void> {
   const app = await NestFactory.create(AppModule)
   app.useGlobalPipes(new ValidationPipe())
 
-  const config = new DocumentBuilder()
-    .setTitle('Composte Application API')
-    .setDescription('The composte application API description')
-    .setVersion('1.0')
-    .addBearerAuth()
-    .build()
-  const document = SwaggerModule.createDocument(app, config)
-  SwaggerModule.setup('doc', app, document)
+  setupSwagger(app)
 
   await app.listen(3000)
 }
