+++ src/swagger.setup.ts
@@ -1,0 +1,13 @@
+import { INestApplication } from '@nestjs/common'
+import { DocumentBuilder, SwaggerModule } from '@nestjs/swagger'
+
+export const setupSwagger = (app: INestApplication): void => {
+  const config = new DocumentBuilder()
+    .setTitle('Composte Application API')
+    .setDescription('The composte application API description')
+    .setVersion('1.0')
+    .addBearerAuth()
+    .build()
+  const document = SwaggerModule.createDocument(app, config)
+  SwaggerModule.setup('doc', app, document)
+}
